+++ libs/cmdb-instances/src/cmdb-instances-input-form/CmdbInstancesInputFormItem.tsx
@@ -20,6 +20,7 @@ export interface CmdbInstancesInputFormItemProps {
   objectMap?: { [objectId: string]: Partial<CmdbModels.ModelCmdbObject> };
   objectId?: string;
   fieldId?: string;
+  fields?: string[];
   singleSelect?: boolean;
   inputDisabled?: boolean;
   checkDisabled?: boolean;
@@ -34,6 +35,9 @@ export interface CmdbInstancesInputFormItemProps {
   defaultQuery?: { [fieldId: string]: any }[];
   addButtonDisabled?: boolean;
   previewEnabled?: boolean;
+  pageSize?: number;
+  pageSizeOptions?: string[];
+  showSizeChanger?: boolean;
 }
 
 export const LegacyCmdbInstancesInputFormItem = (
@@ -57,6 +61,15 @@ export const LegacyCmdbInstancesInputFormItem = (
 
     return { presetQuery, permission };
   }, [props.objectId, props.checkAgentStatus, props.checkPermission]);
+  const fields = useMemo(() => {
+    const fields: Record<string, boolean> = {};
+
+    props.fields?.forEach((field) => {
+      fields[field] = true;
+    });
+
+    return fields;
+  }, [props.fields]);
 
   const objectData = modifyModelData(props.objectMap[props.objectId]);
   const fieldData = objectData.__fieldList.find(
@@ -102,17 +115,19 @@ export const LegacyCmdbInstancesInputFormItem = (
             ...presetQuery,
           },
           permission,
-          fields: {
+          fields: fields || {
             instanceId: true,
             [props.fieldId]: true,
           },
         })
       ).list;
-      const instanceIds = instances.map((instance) => instance.instanceId);
+      const instanceIdSet = new Set(
+        instances.map((instance) => instance.instanceId)
+      );
 
       const selectedInstancesMap = groupBy(
         selectedInstances,
-        (selectedInstance) => instanceIds.includes(selectedInstance.instanceId)
+        (selectedInstance) => instanceIdSet.has(selectedInstance.instanceId)
       );
 
       setSelectedInstances({
@@ -122,13 +137,11 @@ export const LegacyCmdbInstancesInputFormItem = (
         ),
       });
 
-      return selectedInstancesMap.true
-        ? selectedInstancesMap.true.map((instance) => instance.instanceId)
-        : [];
+      return selectedInstancesMap.true ? selectedInstancesMap.true : [];
     }
   };
 
-  const updateSelected = async (instanceIds: string[]): Promise<string[]> => {
+  const updateSelected = async (instanceIds: string[]): Promise<any[]> => {
     let selectedInstances: any[] = [];
     if (instanceIds.length) {
       selectedInstances = (
@@ -140,7 +153,7 @@ export const LegacyCmdbInstancesInputFormItem = (
               $in: instanceIds,
             },
           },
-          fields: {
+          fields: fields || {
             instanceId: true,
             [props.fieldId]: true,
           },
@@ -154,17 +167,15 @@ export const LegacyCmdbInstancesInputFormItem = (
         .join(separator)
     );
 
-    let keys: string[] = [];
+    let instances = selectedInstances;
     if (selectedInstances.length) {
       if (!props.checkDisabled) {
-        keys = await checkSelectedInstances(selectedInstances);
-      } else {
-        keys = selectedInstances.map((instance) => instance.instanceId);
+        instances = await checkSelectedInstances(selectedInstances);
       }
     } else {
       setSelectedInstances({ valid: [], invalid: [] });
     }
-    return keys;
+    return instances;
   };
 
   useEffect(() => {
@@ -173,6 +184,11 @@ export const LegacyCmdbInstancesInputFormItem = (
     }
     // todo(ice): how formItem to change its value??????
   }, []);
+
+  const handleChange = (instances: any[]): void => {
+    props.onChange?.(instances.map((instance) => instance.instanceId));
+    props.onChangeV2?.(instances);
+  };
 
   const checkInputValue = async (inputValue: string): Promise<void> => {
     const fieldValues =
@@ -197,7 +213,7 @@ export const LegacyCmdbInstancesInputFormItem = (
             ],
           },
           permission,
-          fields: {
+          fields: fields || {
             instanceId: true,
             [props.fieldId]: true,
           },
@@ -220,9 +236,7 @@ export const LegacyCmdbInstancesInputFormItem = (
         invalid: invalidFieldValues || [],
       });
 
-      props.onChange?.(
-        validSelectedInstances.map((instance) => instance.instanceId)
-      );
+      handleChange(validSelectedInstances);
     }
   };
 
@@ -239,13 +253,9 @@ export const LegacyCmdbInstancesInputFormItem = (
   ): Promise<void> => {
     closeSelectInstancesModal();
 
-    const keys = await updateSelected(selectedKeys);
+    const instances = await updateSelected(selectedKeys);
 
-    props.onChange?.(keys);
-  };
-
-  const handleInstancesSelectedV2 = (instanceDataList: any[]) => {
-    props.onChangeV2?.(instanceDataList);
+    handleChange(instances);
   };
 
   const handleInputChanged = (event: { target: { value: string } }): void => {
@@ -279,7 +289,7 @@ export const LegacyCmdbInstancesInputFormItem = (
                 ...(props.query ? props.query : []),
               ],
             },
-            fields: {
+            fields: fields || {
               instanceId: true,
               [props.fieldId]: true,
             },
@@ -291,7 +301,7 @@ export const LegacyCmdbInstancesInputFormItem = (
           invalid: [],
         });
 
-        props.onChange?.(instances.map((instance) => instance.instanceId));
+        handleChange(instances);
       }
     }
 
@@ -349,16 +359,18 @@ export const LegacyCmdbInstancesInputFormItem = (
         visible={visible}
         title={text}
         aq={props.query}
-        presetConfigs={{ query: presetQuery }}
+        presetConfigs={{ query: presetQuery, fieldIds: props.fields }}
         permission={permission}
         selectedRowKeys={selectedInstances.valid.map(
           (instance) => instance.instanceId
         )}
         onSelected={handleInstancesSelected}
-        onSelectedV2={handleInstancesSelectedV2}
         singleSelect={props.singleSelect}
         onCancel={closeSelectInstancesModal}
         defaultQuery={props.defaultQuery}
+        pageSize={props.pageSize}
+        pageSizeOptions={props.pageSizeOptions}
+        showSizeChanger={props.showSizeChanger}
         data-testid="select-modal"
       />
       <div className={style.cmdbInstancesInputFormWrapper}>
@@ -387,7 +399,7 @@ export const LegacyCmdbInstancesInputFormItem = (
             visible={previewVisible}
             title={t(K.VIEW_SPECIFIC_INSTANCES)}
             aq={selectedInstancesQuery}
-            presetConfigs={{ query: presetQuery }}
+            presetConfigs={{ query: presetQuery, fieldIds: props.fields }}
             permission={permission}
             onCancel={() => setPreviewVisible(false)}
             defaultQuery={props.defaultQuery}
@@ -397,6 +409,9 @@ export const LegacyCmdbInstancesInputFormItem = (
             relatedToMeDisabled
             moreButtonsDisabled
             selectDisabled
+            pageSize={props.pageSize}
+            pageSizeOptions={props.pageSizeOptions}
+            showSizeChanger={props.showSizeChanger}
             data-testid="preview-modal"
           />
           <Button